# public/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Movie Titles - API",
    "description": "This is the API for the cinema guru application.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/watch-later": {
      "get": {
        "tags": ["Watch Later"],
        "summary": "Get list of movie titles added by logged in user to their watch later list",
        "description": "Get list of movie titles added by logged in user to their watch later list",
        "operationId": "getWatchLater",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Page number",
            "schema": {
              "type": "number"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "titles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MovieTitle"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MovieTitle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "038fddca-61d3-4ccd-be11-f17a7fe39fdc"
          },
          "title": {
            "type": "string",
            "example": "Echo Valley"
          },
          "synopsis": {
            "type": "string",
            "example": "Suspendisse sed facilisis tortor, nec ornare eros. Maecenas ut ullamcorper velit, nec efficitur dui. Nullam vulputate sapien nec tincidunt varius."
          },
          "released": {
            "type": "number",
            "example": "2023"
          },
          "genre": {
            "type": "string",
            "example": "Mystery"
          },
          "favorited": {
            "type": "boolean",
            "example": false
          },
          "watchLater": {
            "type": "boolean",
            "example": true
          },
          "image": {
            "type": "string",
            "example": "/images/038fddca-61d3-4ccd-be11-f17a7fe39fdc.webp"
          }
        }
      }
    }
  }
}